Add helper class to have methods used multiple times
